refactor: chnage NavLink level and remove manaCovector from Component:
