+++ react-magic/src/Card/style.scss
@@ -17,6 +17,7 @@
         width: 100%;
         height: 100%;
         color: white;
+        text-decoration: none;
         background: rgba(0, 0, 0, .7);
         transition: all 1s;
