+++ react-magic/src/CardDetail/CardDetail.tsx
@@ -4,6 +4,7 @@ import Card from '../Card/Card';
 import Loader from '../Loader/Loader';
 import { ICardInfo } from '../types';
 import './style.scss';
+import { manaCostCoverter } from './utils';
 
 interface ICardPageParams {
   cardID: string;
@@ -14,39 +15,6 @@ export default function CardDetail(): JSX.Element {
   const [isDownloaded, setDownloading] = useState(false);
   const [cardInfo, setInfo] = useState<ICardInfo>({});
 
-  function manaCostCoverter(manaString: string | undefined): JSX.Element [] {
-    if (!manaString) return [];
-    const manaArr = manaString.split(/[}{]/).filter((str) => !!str);
-    const jsxArr = [];
-    for (let i = 0; i < manaArr.length; i++) {
-      if (manaArr[i].match(/\d/)) {
-        jsxArr.push(<p className="mana mana_number" key={Math.random() * 99999999}>{manaArr[i]}</p>);
-      } else {
-        let path: string;
-        switch (manaArr[i]) {
-          case 'W':
-            path = 'white.png';
-            break;
-          case 'G':
-            path = 'green.png';
-            break;
-          case 'R':
-            path = 'red.png';
-            break;
-          case 'U':
-            path = 'blue.png';
-            break;
-          case 'B':
-            path = 'black.png';
-            break;
-          default:
-            path = '';
-        }
-        jsxArr.push(<img className="mana mana_color" key={Math.random() * 99999999} src={`./${path}`}/>);
-      }
-    }
-    return jsxArr;
-  }
   async function getUser(): Promise<void> {
     try {
       const req = await fetch(`https://api.magicthegathering.io/v1/cards?id=${cardID}`);
